ABSTRACT
The main objective of the research is to clarify the significant differences between Industry 4.0 and Industry 5.0 and to demonstrate why the latter is the superior option. In its infancy, the Industry  
4.0 phenomenon poses major obstacles to all existing business models and accelerates the need to rethink and digitize operations. The literature on supply chain management was also offered, but as we will see, the Smart supply chain is not advancing the processing of industry 4.0 and its capabilities are in their infancy. However, the relatively recent phenomenon of Industry 5.0 seems to be a more accurate, efficient, and time-saving option for the future of supply chain management in a globalized world. To gather data from the existing body of knowledge on this topic, a systematic literature review methodology was used. The results have shown that the supply chain sector is unprepared for Industry 5.0 approaches.  
Based on the essential characteristics of Industry 4.0, the supply chain, and the growing body of literature addressing disruptions to Smart SCs, this research employs a narrative literature approach to provide an integrative framework to shed light on Smart SCs and their blind spots. The four (4) pillars of Industry 5.0 industry strategy, innovation and technologies, society and sustainability, and transition issues were recommended to serve as the basis for the emerging Supply Chain 5.0 framework and accompanying research agenda. Industry 5.0 is still in its infancy and in a great position. Due to the paucity of relevant literature, it is anticipated that this study will generate an abundance of new ideas and conclusions.